+++ buildVars.py
@@ -23,11 +23,13 @@ addon_info = {
 	"addon_summary": _("Visor para las consolas de Windows"),
 	# Add-on description
 	# Translators: Long description to be shown for this add-on on add-on information from add-ons manager
-	"addon_description": _("""Pequeño complemento para mostrar en un dialogo el contenido de las consolas de Windows
+	"addon_description": _("""* Pequeño complemento para mostrar en un dialogo el contenido de las consolas de Windows
+* Lanzador de consolas CMD, PowerShell, Terminal, Bash, Visual Studio developer.
+* Copiado del portapapeles en consola enfocada de manera fácil.
 
-Asignar tecla en Gestos de entrada / Visor de consola."""),
+Asignar teclas en Gestos de entrada / Visor de consola."""),
 	# version
-	"addon_version": "1.3",
+	"addon_version": "1.4",
 	# Author(s)
 	"addon_author": u"Héctor J. Benítez Corredera <xebolax@gmail.com>",
 	# URL for the add-on documentation support
@@ -39,7 +41,7 @@ Asignar tecla en Gestos de entrada / Visor de consola."""),
 	# Minimum NVDA version supported (e.g. "2018.3.0", minor version is optional)
 	"addon_minimumNVDAVersion": "2022.1.0",
 	# Last NVDA version supported/tested (e.g. "2018.4.0", ideally more recent than minimum version)
-	"addon_lastTestedNVDAVersion": "2024.1.0",
+	"addon_lastTestedNVDAVersion": "2024.4.1",
 	# Add-on update channel (default is None, denoting stable releases,
 	# and for development releases, use "dev".)
 	# Do not change unless you know what you are doing!