+++ changelog.md
@@ -1,1 +1,2 @@
-* Agregada compatibilidad con Windows Terminal.
+* Lanzador de consolas CMD, PowerShell, Terminal, Bash, Visual Studio developer.
+* Copiado del portapapeles en consola enfocada de manera fácil.